refactor(ai-search-launcher): extract ChatWindow to src/common/components/chat
