+++ src/common/components/chat/constants.ts
